Add template-bind
The extension of the template using a bind attribute that specifies the
class.

It actually tries to bind properties and methods on a given object to
the active polymer element. It is likely that this is incorrect so it
may require that the polymer element be constructed in code and
correctly bound at that point.